tests: Timeout stuck tests
The partial go implementation cause the tests to hang. Add global and
specific tests timeouts to to make the test fail when the program does
not respond fast enough.

This requires now the pytest-timeout package. Add a standard pip
requirements.txt file, so travis will install the needed packages, and
update the readme.