go: Improve the monitoring loop
Previously if a monitor was blocked on storage, ticks were ignored. Now
the actual checking is done in a separate goroutine, and when the ticker
ticks while the last check is blocking we log a warning about the
blocked monitor.

If the monitor was stopped while blocked on storage, the monitor would
see the stop request only when the check completed. Now the monitor
receive the request immediately, and is sending an EINPROGRESS response.

Because the monitor loop is not blocking now, we can use a blocking
channel for the stop channel, and we don't need to do non-blocking
send.

This go version behaves now like the C version.